Made deleteJunction the only operation possible on a junction, even when selecting multiple at a time. Fixes at least one bug and removes ambiguous behavior.
This isn't entirely intuitive when selecting both types and pressing delete but it doesn't create any problems by unexpectedly deleting things. A single button to delete both a junction and its target will be implemented in the single pane view which will probably supersede any desire to have it here.