Handle abnormally named junction points (compact interface) Allow Archive on the compact interface to create junction points for entries that only exist in the destination.
Junction points with names different from their target are now not displayed on the compact interface because that is not an officially supported scenario and there is no satisfactory method to display multiple junction points with the same target. This does not affect behavior of the commands (delete will still remove all junction points, etc), only what is displayed on the compact interface.

Related Work Items: #8