Display how many items would be affected by each action
Added badges to each button for how many items apply to that command and removed the TextBlock counter of number of selected items.

Changed value converters to use a MarkupExtension instead of requiring a resource.
Fixed Restore throwing an error after navigating away and then back to a directory mapping.