Fix NameFilter typing lag
Adding and removing live filtering properties is currently an expensive operation as bindings need to be created for each item so the live filtering properties were made constant. The extra work created by tracking properties that don't affect the filter has no noticeable performance degredation. Names do not need to be included in the LiveFiltering because an earlier commit changed names to never be modified in place- renaming causes items to be removed and then re-added.

Related Work Items: #25